+++ dymension/libs/kaspa/lib/relayer/src/deposit.rs
@@ -1,9 +1,18 @@
+use std::error::Error;
+
 use crate::handle_new_deposit;
-use corelib::api::deposits::Deposit;
 use corelib::deposit::DepositFXG;
 use eyre::Result;
 
-pub async fn on_new_deposit(deposit: &Deposit) -> Result<Option<DepositFXG>> {
-    let deposit_tx_result = handle_new_deposit(deposit.id.to_string()).await?;
-    Ok(Some(deposit_tx_result))
+pub async fn handle_new_deposits(
+    transaction_ids: Vec<String>,
+) -> Result<Vec<DepositFXG>, Box<dyn Error>> {
+    let mut txs = Vec::new();
+
+    for transaction in transaction_ids {
+        let tx = handle_new_deposit(transaction).await?;
+        txs.push(tx);
+    }
+
+    Ok(txs)
 }
