+++ dymension/libs/kaspa/lib/relayer/src/lib.rs
@@ -123,19 +123,6 @@ pub async fn handle_new_deposit(tx: String) -> Result<DepositFXG> {
     Ok(tx)
 }
 
-pub async fn handle_new_deposits(
-    transaction_ids: Vec<String>,
-) -> Result<Vec<DepositFXG>, Box<dyn Error>> {
-    let mut txs = Vec::new();
-
-    for transaction in transaction_ids {
-        let tx = handle_new_deposit(transaction).await?;
-        txs.push(tx);
-    }
-
-    Ok(txs)
-}
-
 #[cfg(test)]
 mod tests {
     use hyperlane_core::{Encode, H256};
